chore: rename _components folder name
