+++ screens/gallery/galleryComponents/AlbumCard.js
